Solution to Count Leading Zeros in a Binary Number

+++ Count Leading Zeros in a Binary Number@@67882a731b4392d8fc1577f2/code.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 int main() {
-    long long int n;
-    scanf("%lld",&n);
-    long long int k = 1,count=0;
+    long int n;
+    scanf("%ld",&n);
+    long int k = 1, count=0;
     while(k<n){
         k = k<<1;
         count++;
     }
-    printf("%lld",31-count);
+    printf("%ld",31-count);
     return 0;
 }
